Move admin links from components endpoint into a new singleton instance
